Extract Google Pay availability check into new class
So the logic can be shared between the old and new implementation.

COAND-943
